Add files via upload
This D-Mart Shopping Website is a simple e-commerce website built using Bootstrap, designed to emulate the experience of shopping at an online supermarket. The project showcases the use of Bootstrap’s responsive grid system and components to create a clean, modern, and mobile-friendly layout, which includes features like product listings, shopping cart, and basic navigation. The aim is to provide a seamless user experience while shopping for groceries, daily essentials, and other products on the platform.